% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin.test}
\alias{bin.test}
\title{Function Specific for Binary Simulation}
\usage{
bin.test(X, k, F.init, G.init, method, prob, iter = 500, tol = 1e-05,
  tau = 0.5, step_bin = 0.05, step_log = 0.001, factor = 2)
}
\description{
Function to apply various methods of NMF on the input matrix for both continuous and binary entries.
}
