% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{error.mess}
\alias{error.mess}
\title{Error function}
\usage{
error.mess(X, mode, k, method, init, iter, tol, tau, step_bin, step_log,
  seed)
}
\description{
Outputs error and warning messages
This function examines the input parameters and output corresponding error or warning message.
}
