\name{MatrixFact-package}
\alias{MatrixFact-package}
\alias{MatrixFact}
\docType{package}
\title{
\packageTitle{MatrixFact}
}
\description{
\packageDescription{MatrixFact}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MatrixFact}
\packageIndices{MatrixFact}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MatrixFact}

Maintainer: \packageMaintainer{MatrixFact}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
