% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bin.test}
\alias{bin.test}
\title{Function Specific for Binary Simulation}
\usage{
bin.test(X, k, F.init, G.init, method, prob, iter = 500, tol = 1e-05,
  tau = 0.5, step_bin = 0.05, step_log = 0.001, factor = 2)
}
\description{
This function is for simulation purposes only.
}
\details{
This function is the same as nmf.main(), with the additional parameter of an input
true probability matrix for simulation purpose. This is not applicable in actual
scenarios, since there is no way of knowing the true underlying probability matrix
beforehand.
}
